Added tab component and breadcrumb routing display

+++ src/components/admin/adminHeader.vue
@@ -1,17 +1,18 @@
 <script setup lang="ts">
 import fullScreen from "@/components/fullScreen";
 import theme from "@/components/theme.js";
-import {ArrowDown} from "@element-plus/icons-vue";
-import {useRouter} from "vue-router";
-import {useAuthStore} from "@/stores/auth";
-import {getUserInfo, logout} from "@/api";
-import {onMounted} from "vue";
-import {ElMessage} from "element-plus";
-import type {UserInfo} from "@/types/define";
-import {errorCode} from "@/utils/errcode";
+import { ArrowDown } from "@element-plus/icons-vue";
+import { useRoute, useRouter } from "vue-router";
+import { useAuthStore } from "@/stores/auth";
+import { getUserInfo, logout } from "@/api";
+import { onMounted } from "vue";
+import { ElMessage } from "element-plus";
+import type { UserInfo } from "@/types";
+import { errorCode } from "@/utils/errcode";
 
 const authStore = useAuthStore();
 const router = useRouter();
+const route = useRoute();
 let userInfo = reactive<UserInfo>({
   id: 0,
   email: "",
@@ -21,12 +22,12 @@ let userInfo = reactive<UserInfo>({
   sex: "",
   avatar: "",
   nick_name: "",
-})
+});
 
 function returnIndex() {
   router.push({
-    path: "/"
-  })
+    path: "/",
+  });
 }
 
 // 处理错误逻辑
@@ -42,7 +43,10 @@ const handleError = (error: any) => {
   }
 };
 
-const showMessage = (message: string, type: 'success' | 'warning' | 'info' | 'error') => {
+const showMessage = (
+  message: string,
+  type: "success" | "warning" | "info" | "error"
+) => {
   ElMessage({
     message,
     type,
@@ -59,44 +63,71 @@ const receiveUserInfo = async () => {
       authStore.setCheckTokenData(userInfo); // 存储用户信息到状态管理器
     }
   } catch (error: any) {
-    handleError(error)
+    handleError(error);
   }
 };
 
 // 登出
 const toLogout = async () => {
   try {
-    const res = await logout(); // 假设 logout 方法会处理请求并返回响应
-    // 检查响应状态或数据，根据需要进行处理
-    if (res.data.status === 0) {
-      showMessage(res.data.message || '登出成功', "success")
-      authStore.deleteToken() // 清除 token
-      await router.push({name: 'login'}); // 重定向到登录页面
+    const confirm = await ElMessageBox.confirm("确定要登出吗？", "提示", {
+      confirmButtonText: "确认",
+      cancelButtonText: "取消",
+      type: "warning",
+    });
+
+    if (confirm) {
+      const res = await logout(); // 假设 logout 方法会处理请求并返回响应
+      // 检查响应状态或数据，根据需要进行处理
+      if (res.data.status === 0) {
+        showMessage(res.data.message || "登出成功", "success");
+        authStore.deleteToken(); // 清除 token
+        await router.push({ name: "login" }); // 重定向到登录页面
+      }
     }
   } catch (error: any) {
-    handleError(error)
+    // 如果是用户取消操作，不处理
+    if (error !== "cancel") {
+      handleError(error);
+    }
   }
 };
+  // 用于面包屑
+const breadcrumbRoutes = computed(() => {
+  return route.matched.filter(item => item.meta && item.meta.title);
+});
 onMounted(() => {
   receiveUserInfo();
-})
+
+});
 </script>
 
 <template>
-  <div class=" flex justify-between items-center  h-16">
+  <div class="flex justify-between items-center h-16">
     <div class="header-left flex items-center">
-      <img src="https://element-plus.org/images/element-plus-logo.svg" alt="Logo" class="h-10">
+      <img
+        src="https://element-plus.org/images/element-plus-logo.svg"
+        alt="Logo"
+        class="h-10"
+      />
       <!-- 头部面包屑 -->
-      <el-breadcrumb class="ml-4 text-base outline-none">
-        <el-breadcrumb-item :to="{ path: '/' }">首页</el-breadcrumb-item>
-        <el-breadcrumb-item>
-          <a href="/" class="text-blue-600">用户管理</a>
+      <el-breadcrumb class="ml-6" >
+        <el-breadcrumb-item
+          v-for="item in breadcrumbRoutes"
+          :key="item.name"
+          :to="{ name: item?.name }"
+          class="text-base"
+        >
+          {{ item?.meta?.title }}
         </el-breadcrumb-item>
       </el-breadcrumb>
     </div>
     <div class="header-right flex items-center">
       <el-tooltip content="返回前台首页">
-        <i class="iconfont icon-Home text-3xl cursor-pointer mr-2.5" @click="returnIndex"></i>
+        <i
+          class="iconfont icon-Home text-3xl cursor-pointer mr-2.5"
+          @click="returnIndex"
+        ></i>
       </el-tooltip>
       <!-- 全屏组件 -->
       <fullScreen class="mr-2.5"></fullScreen>
@@ -105,12 +136,13 @@ onMounted(() => {
         <theme class="mr-3"></theme>
       </el-tooltip>
       <div class="avatar mr-2.5">
-        <el-avatar :src="userInfo.avatar"/>
+        <el-avatar :src="userInfo.avatar" />
       </div>
       <!-- 下拉菜单 -->
       <el-dropdown placement="bottom-end">
-        <span class=" w-full flex items-center cursor-pointer outline-none">
-          {{ userInfo.name }} <el-icon size="12px" class="ml-1"><ArrowDown/></el-icon>
+        <span class="w-full flex items-center cursor-pointer outline-none">
+          {{ userInfo.name }}
+          <el-icon size="12px" class="ml-1"><ArrowDown /></el-icon>
         </span>
         <template #dropdown>
           <el-dropdown-menu>
@@ -123,6 +155,4 @@ onMounted(() => {
     </div>
   </div>
 </template>
-<style lang="scss" scoped>
-
-</style>
+<style lang="scss" scoped></style>
